Resolve package splitting geode-membership
